[CALCITE-5909] SqlParserTest.testNoUnintendedNewReservedKeywords fails in IDE while passes in command line
