+++ testkit/src/main/java/org/apache/calcite/sql/parser/SqlParserTest.java
@@ -38,7 +38,6 @@ import org.apache.calcite.sql.type.SqlTypeName;
 import org.apache.calcite.sql.util.SqlShuttle;
 import org.apache.calcite.sql.validate.SqlConformance;
 import org.apache.calcite.sql.validate.SqlConformanceEnum;
-import org.apache.calcite.test.DiffTestCase;
 import org.apache.calcite.test.IntervalTest;
 import org.apache.calcite.tools.Hoist;
 import org.apache.calcite.util.Bug;
@@ -48,7 +47,6 @@ import org.apache.calcite.util.TestUtil;
 import org.apache.calcite.util.Util;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSortedSet;
 
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -66,7 +64,6 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Random;
 import java.util.SortedSet;
-import java.util.TreeSet;
 import java.util.function.Consumer;
 import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
@@ -85,7 +82,6 @@ import static org.hamcrest.Matchers.hasToString;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotSame;
 import static org.junit.jupiter.api.Assumptions.assumeFalse;
-import static org.junit.jupiter.api.Assumptions.assumeTrue;
 
 /**
  * A <code>SqlParserTest</code> is a unit-test for
@@ -107,7 +103,7 @@ public class SqlParserTest {
    * the SQL:92, SQL:99, SQL:2003, SQL:2011, SQL:2014 standards and in Calcite.
    *
    * <p>The standard keywords are derived from
-   * <a href="https://developer.mimer.com/wp-content/uploads/2018/05/Standard-SQL-Reserved-Words-Summary.pdf">Mimer</a>
+   * <a href="https://developer.mimer.com/wp-content/uploads/standard-sql-reserved-words-summary.pdf">Mimer</a>
    * and from the specification.
    *
    * <p>If a new <b>reserved</b> keyword is added to the parser, include it in
@@ -325,6 +321,7 @@ public class SqlParserTest {
       "JSON_OBJECT",                                                       "c",
       "JSON_OBJECTAGG",                                                    "c",
       "JSON_QUERY",                                                        "c",
+      "JSON_SCOPE",                                                        "c",
       "JSON_VALUE",                                                        "c",
       "KEEP",                                              "2011",
       "KEY",                           "92", "99",
@@ -400,6 +397,7 @@ public class SqlParserTest {
       "OPTION",                        "92", "99",
       "OR",                            "92", "99", "2003", "2011", "2014", "c",
       "ORDER",                         "92", "99", "2003", "2011", "2014", "c",
+      "ORDINAL",                                                           "c",
       "ORDINALITY",                          "99",
       "OUT",                           "92", "99", "2003", "2011", "2014", "c",
       "OUTER",                         "92", "99", "2003", "2011", "2014", "c",
@@ -434,6 +432,7 @@ public class SqlParserTest {
       "PRIVILEGES",                    "92", "99",
       "PROCEDURE",                     "92", "99", "2003", "2011", "2014", "c",
       "PUBLIC",                        "92", "99",
+      "QUALIFY",                                                           "c",
       "RANGE",                               "99", "2003", "2011", "2014", "c",
       "RANK",                                              "2011", "2014", "c",
       "READ",                          "92", "99",
@@ -472,6 +471,9 @@ public class SqlParserTest {
       "ROWS",                          "92", "99", "2003", "2011", "2014", "c",
       "ROW_NUMBER",                                        "2011", "2014", "c",
       "RUNNING",                                                   "2014", "c",
+      "SAFE_CAST",                                                         "c",
+      "SAFE_OFFSET",                                                       "c",
+      "SAFE_ORDINAL",                                                      "c",
       "SATURDAY",                                                          "c",
       "SAVEPOINT",                           "99", "2003", "2011", "2014", "c",
       "SCHEMA",                        "92", "99",
@@ -544,6 +546,7 @@ public class SqlParserTest {
       "TRIM_ARRAY",                                        "2011", "2014", "c",
       "TRUE",                          "92", "99", "2003", "2011", "2014", "c",
       "TRUNCATE",                                          "2011", "2014", "c",
+      "TRY_CAST",                                                          "c",
       "TUESDAY",                                                           "c",
       "UESCAPE",                                           "2011", "2014", "c",
       "UNDER",                               "99",
@@ -7297,39 +7300,6 @@ public class SqlParserTest {
     assertThat(!jdbcKeywords.contains(",SELECT,"), is(true));
   }
 
-  /**
-   * Tests that reserved keywords are not added to the parser unintentionally.
-   * (Most keywords are non-reserved. The set of reserved words generally
-   * only changes with a new version of the SQL standard.)
-   *
-   * <p>If the new keyword added is intended to be a reserved keyword, update
-   * the {@link #RESERVED_KEYWORDS} list. If not, add the keyword to the
-   * non-reserved keyword list in the parser.
-   */
-  @Test void testNoUnintendedNewReservedKeywords() {
-    assumeTrue(isNotSubclass(), "don't run this test for sub-classes");
-    final SqlAbstractParserImpl.Metadata metadata =
-        fixture().parser().getMetadata();
-
-    final SortedSet<String> reservedKeywords = new TreeSet<>();
-    final SortedSet<String> keywords92 = keywords("92");
-    for (String s : metadata.getTokens()) {
-      if (metadata.isKeyword(s) && metadata.isReservedWord(s)) {
-        reservedKeywords.add(s);
-      }
-      // Check that the parser's list of SQL:92
-      // reserved words is consistent with keywords("92").
-      assertThat(s, metadata.isSql92ReservedWord(s),
-          is(keywords92.contains(s)));
-    }
-
-    final String reason = "The parser has at least one new reserved keyword. "
-        + "Are you sure it should be reserved? Difference:\n"
-        + DiffTestCase.diffLines(ImmutableList.copyOf(getReservedKeywords()),
-        ImmutableList.copyOf(reservedKeywords));
-    assertThat(reason, reservedKeywords, is(getReservedKeywords()));
-  }
-
   @Test void testTabStop() {
     sql("SELECT *\n\tFROM mytable")
         .ok("SELECT *\n"
@@ -9282,10 +9252,6 @@ public class SqlParserTest {
       SqlTests.checkEx(thrown, expectedMsgPattern, sap,
           SqlTests.Stage.VALIDATE);
     }
-  }
-
-  private boolean isNotSubclass() {
-    return this.getClass().equals(SqlParserTest.class);
   }
 
   /**
